+++ src/command_modules/azure-cli-cdn/azure/cli/command_modules/cdn/tests/latest/test_endpoint_scenarios.py
@@ -6,12 +6,12 @@ from azure.cli.testsdk import ResourceGroupPreparer, JMESPathCheck
 from azure.cli.testsdk import ScenarioTest
 from .scenario_mixin import CdnScenarioMixin
 
+from azure.mgmt.cdn.models import SkuName
+
 
 class CdnEndpointScenarioTest(CdnScenarioMixin, ScenarioTest):
     @ResourceGroupPreparer()
     def test_endpoint_crud(self, resource_group):
-        from knack.util import CLIError
-
         profile_name = 'profile123'
         self.endpoint_list_cmd(resource_group, profile_name, expect_failure=True)
 
@@ -91,3 +91,34 @@ class CdnEndpointScenarioTest(CdnScenarioMixin, ScenarioTest):
 
         content_paths = ['/index.html', '/javascript/*']
         self.endpoint_purge_cmd(resource_group, endpoint_name, profile_name, content_paths)
+
+    @ResourceGroupPreparer()
+    def test_endpoint_different_profiles(self, resource_group):
+        origin = 'www.example.com'
+        checks = [JMESPathCheck('origins[0].hostName', origin),
+                  JMESPathCheck('isHttpAllowed', True),
+                  JMESPathCheck('isHttpsAllowed', True),
+                  JMESPathCheck('isCompressionEnabled', False),
+                  JMESPathCheck('queryStringCachingBehavior', 'IgnoreQueryString')]
+
+        # create an endpoint using the standard_akamai profile
+        profile_name = self._create_profile(resource_group, SkuName.standard_akamai.value)
+        endpoint_name = self.create_random_name(prefix='endpoint', length=24)
+        self.endpoint_create_cmd(resource_group, endpoint_name, profile_name, origin, checks=checks + [JMESPathCheck('name', endpoint_name)])
+
+        # create an endpoint using the standard_verizon profile
+        profile_name = self._create_profile(resource_group, SkuName.standard_verizon.value)
+        endpoint_name = self.create_random_name(prefix='endpoint', length=24)
+        self.endpoint_create_cmd(resource_group, endpoint_name, profile_name, origin, checks=checks + [JMESPathCheck('name', endpoint_name)])
+
+        # create an endpoint using the premium_verizon profile
+        profile_name = self._create_profile(resource_group, SkuName.premium_verizon.value)
+        endpoint_name = self.create_random_name(prefix='endpoint', length=24)
+        premium_checks = checks.extend([JMESPathCheck('queryStringCachingBehavior', 'NotSet'), JMESPathCheck('name', endpoint_name)])
+        self.endpoint_create_cmd(resource_group, endpoint_name, profile_name, origin, checks=premium_checks)
+
+    def _create_profile(self, resource_group, profile_sku):
+        profile_name = (profile_sku + "_profile").replace("_", "-")  # profile names must match ^[a-zA-Z0-9]+(-*[a-zA-Z0-9])*$
+        check = JMESPathCheck('sku.name', profile_sku)
+        self.profile_create_cmd(resource_group, profile_name, sku=profile_sku, checks=check)
+        return profile_name