+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/_help.py
@@ -100,6 +100,13 @@ helps['storage container create'] = """
           text: az storage container create -n MyStorageContainer --fail-on-exist
 """
 
+helps['storage container delete'] = """
+    type: command
+    short-summary: Marks the specified container for deletion.
+    long-summary: >
+        The container and any blobs contained within it are later deleted during garbage collection.
+"""
+
 helps['storage account list'] = """
     type: command
     short-summary: List storage accounts.
@@ -743,6 +750,11 @@ helps['storage queue policy'] = """
 helps['storage share'] = """
     type: group
     short-summary: Manage file shares.
+"""
+
+helps['storage share url'] = """
+    type: command
+    short-summary: Create a URI to access a file share.
 """
 
 helps['storage share exists'] = """