+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ===============
 
+2.3.9
++++++
+* Minor fixes
+
+2.3.8
++++++
+* Minor fixes.
+
 2.3.7
 +++++
 * Minor fixes.