+++ src/command_modules/azure-cli-container/HISTORY.rst
@@ -3,6 +3,19 @@
 Release History
 ===============
 
+0.3.7
++++++
+* Make 'Private' a valid type to pass to '--ip-address'
+* Allow using only subnet ID to setup a virtual network for the container group
+* Allow using vnet name or resource id to enable using vnets from different resource groups
+
+0.3.6
++++++
+* Add '--assign-identity' for adding a MSI identity to a container group
+* Add '--scope' to create a role assignment for the system assigned MSI identity
+* Show warning when creating a container group with an image without a long running process
+* Fix table output issues for 'list' and 'show' commands
+
 0.3.5
 +++++
 * Minor changes