+++ src/command_modules/azure-cli-lab/HISTORY.rst
@@ -2,6 +2,9 @@
 
 Release History
 ===============
+0.1.2
++++++
+* Minor changes
 
 0.1.1
 +++++