+++ src/command_modules/azure-cli-monitor/HISTORY.rst
@@ -3,6 +3,19 @@
 Release History
 ===============
 
+0.2.5
++++++
+* `monitor activity-log list`:
+    Allow listing all events at the subscription level.
+    Added `--offset` parameter to more easily create time queries.
+    Improved validation for --start-time and --end-time to use wider range of ISO8601 formats and more user-friendly datetime formats.
+    Added `--namespace` as alias for deprecated option `--resource-provider`.
+    Deprecated `--filters` because no values other than those with strongly-typed options are supported by the service.
+* `monitor metrics list`:
+    Added `--offset` parameter to more easily create time queries.
+    Improved validation for --start-time and --end-time to use wider range of ISO8601 formats and more user-friendly datetime formats.
+* `monitor diagnostic-settings create`: Improve validation for arguments `--event-hub` and `--event-hub-rule`.
+
 0.2.4
 +++++
 * Minor fixes